+++ src/features/SurveyExecution/components/OverruleResult.tsx
@@ -3,11 +3,15 @@ import useModalVisibility from '../../../hooks/useModalVisibility';
 import {CommentType} from '../../../interfaces/standard';
 import TouchableText from '../../../components/TouchableText';
 import CommentModal from '../../../components/CommentModal';
-import {useDispatch} from '../../../utils/store/configureStore';
-import {overruleStandardResult} from '../evaluationReducer';
+import {useDispatch, useSelector} from '../../../utils/store/configureStore';
+import {
+  overruleStandardResult,
+  resetOverruleStandardResult,
+} from '../evaluationReducer';
 import format from 'date-fns/format';
 import {OverruleStatus} from '../../../interfaces/common';
 import {useTranslation} from 'react-i18next';
+import {Button} from 'react-native-paper';
 
 interface OverruleResultProps {
   surveyId: string;
@@ -27,6 +31,11 @@ const chips: Array<{title: string; value: OverruleStatus}> = [
 
 function OverruleResult({surveyId, standardId}: OverruleResultProps) {
   const [visible, handleVisible] = useModalVisibility();
+  const data = useSelector(
+    state =>
+      state.evaluation[surveyId].standards?.find(i => i.id === standardId)
+        ?.overruleComment,
+  );
   const dispatch = useDispatch();
   const handleSave = React.useCallback(
     ({text, chip}: {text: string; chip: CommentType | OverruleStatus}) => {
@@ -42,6 +51,10 @@ function OverruleResult({surveyId, standardId}: OverruleResultProps) {
     },
     [dispatch, standardId, surveyId],
   );
+  const handleResetOverrule = React.useCallback(() => {
+    handleVisible();
+    dispatch(resetOverruleStandardResult({surveyId, standardId}));
+  }, [handleVisible, dispatch, surveyId, standardId]);
   const {t} = useTranslation();
   return (
     <>
@@ -53,6 +66,13 @@ function OverruleResult({surveyId, standardId}: OverruleResultProps) {
           chips={chips}
           validationMessage={t('Fill comment and choose a status')}
           titleText={t('Overrule result')}
+          defaultText={data?.value ?? undefined}
+          defaultChip={getDefaultChip(data?.overruledHint ?? undefined)}
+          extraButtons={[
+            <Button mode="text" onPress={handleResetOverrule}>
+              {t('Reset Overrule')}
+            </Button>,
+          ]}
         />
       )}
       <TouchableText onPress={handleVisible} size="Button">
@@ -63,3 +83,16 @@ function OverruleResult({surveyId, standardId}: OverruleResultProps) {
 }
 
 export default React.memo(OverruleResult);
+
+function getDefaultChip(s?: string): OverruleStatus | undefined {
+  if (!s) {
+    return undefined;
+  }
+  if (s.indexOf('failed') > -1) {
+    return chips[1].value;
+  }
+  if (s.indexOf('passed') > -1) {
+    return chips[0].value;
+  }
+  return undefined;
+}