+++ src/interfaces/filters.ts
@@ -2,5 +2,5 @@ export type FilterValues = Array<FilterItem>;
 
 export interface FilterItem {
   fieldName: string;
-  value: string;
+  value: string | string[];
 }